[FIX] input값 int -> long 으로 타입변경

+++ calculator/src/Utility/DataProcessing.java
@@ -4,7 +4,7 @@ import java.text.NumberFormat;
 
 public class DataProcessing 
 {
-	public String formatNumber(int number) 
+	public String formatNumber(long number) 
 	{
 		NumberFormat numberFormat = NumberFormat.getNumberInstance();
 		return numberFormat.format(number);
